add the prediction module and main model training module
Accounted for edge cases where only one condition exists since the prediction module was failing due to missing keys. I solved this by creating a separate mapping dictionary here:
pipeline_data['one_outcome_mapping_dict']
and using it to check for these cases, then setting the metric to INF for those cases and mapping the correct next classification prediction name.
Note that the mapping takes the form of a tuple mapped to a tuple where the first key tuple is in the same global key format, while the other tuple is the column name and the classification name/string for that one-to-one mapping case. This makes it easier for handling later.